Improve display of error state
Add severity to FreeBuffers, FramesTimedOut and FramesDropped
Remove FPClearErrors from driver and define as dynamic parameter
Implement cycling through errors when there is more than one on the FP

+++ odinApp/src/OdinDataDriver.h
@@ -77,7 +77,7 @@ class OdinDataDriver : public OdinClient
  private:
   char mHostname[512];
   OdinDataRestAPI mAPI;
-  int mODCount;
+  size_t mODCount;
   std::string mDatasetName;
   std::string mProcessPluginName;
